Safari-609.1.11.1/LayoutTests/fast/canvas/webgl/toDataURL-unpremultipliedAlpha-expected.txt

Test toDataURL on a non-premultipledAlpha WebGL context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl.getContextAttributes().premultipledAlpha is undefined.
PASS gl.getContextAttributes().preserveDrawingBuffer is true
PASS getError was expected value: NO_ERROR : Should be no errors from setup.

Fill the WebGL canvas with solid white at 50% transparency.
PASS getError was expected value: NO_ERROR : Should be no errors from drawing.
Convert to a JPEG data URL.
PASS imageUrl && imageUrl.length > 0 is true
Append the image to the document.
PASS successfullyParsed is true

TEST COMPLETE
